+++ dbms/src/Interpreters/QueryLog.h
@@ -60,7 +60,7 @@ struct QueryLogElement
 
     ClientInfo client_info;
 
-    std::vector<UInt32> thread_ids;
+    std::vector<UInt64> thread_ids;
     std::shared_ptr<ProfileEvents::Counters> profile_counters;
     std::shared_ptr<Settings> query_settings;
